arcgisplaces: Search for POIs using ArcGIS 'Places Service'

The ArcGIS 'Places service' is a ready-to-use location service that can search for businesses and geographic locations around the world. It allows you to find, locate, and discover detailed information about each place. Query for places near a point, within a bounding box, filter based on categories, or provide search text. 'arcgisplaces' integrates with 'sf' for out of the box compatibility with other spatial libraries. Learn more in the 'Places service' API reference https://developers.arcgis.com/rest/places/.

Author

Maintainer: Josiah Parry josiah.parry@gmail.com (ORCID)