Search for places near a point by radius

Finds places that are within a given radius of a specified location. The returned places contain basic data such as name, category and location.

Usage

near_point(
  x,
  y,
  radius = 1000,
  search_text = NULL,
  category_id = NULL,
  icon = NULL,
  token = arc_token()
)

Arguments

x

The x coordinate, or longitude, to search from, in WGS84 decimal degrees.

y

The y coordinate, or latitude, to search from, in WGS84 decimal degrees.

radius

Default 1000. The radius in meters to search for places. Maximum value of 10000.

search_text

Default NULL. Free search text for places against names, categories etc. Must be a scalar value.

category_id

Default NULL. A character vector which filters places to those that match the category IDs.

icon

Default NULL. Must be one of "svg", "png" "cim". Determines whether icons are returned and the type of icon to use with a place or category.

token

an object of class httr2_token as generated by auth_code() or related function

Value

An sf object with columns

place_id: The unique Id of this place. The ID can be passed to place_details() to retrieve additional details.

name: The name of the place, or point of interest. You can search for places by name using the searchText property

distance: A double vector of the distance, in meters, from the place to the search point.

categories: A data.frame with two columns category_id and label. Categories are uniquely identified by a categoryId. For example, 17119 identifies a "Bicycle Store" and 10051 identifies a "Stadium". Note that a single place can belong to multiple categories (for example, a petrol station could also have a super-market).

icon: A character vector of the URL for an icon for this place or category in either svg, cim or png format.

geometry: an sfc_POINT object in EPSG:4326

References

API Documentation

Examples

if (FALSE) { # \dontrun{
near_point(-117.194769, 34.057289)
near_point(139.75, 35.66)
} # }